rework replication approach & add all-time flag
update the process for handling sharded files to have sling load each file individually
Prior implementation replicated all files in the bucket at once which resulted in slow
performance with a large amount of files.

Added an option to sync all of the available files for a table. Useful for intraday.